Do not wait for debugger to start running sketch
By default, the java VM is started with options for attaching a remote
debugger.  The sketch is suspended until the remote debugger connects.
This always succeeds the first time a sketch is run.  At least on OS X
10.9, this seems to be very fragile, and successive runs of the sketch
often fail to start.  This commit tells the VM to *not* wait for the
debugger before starting the sketch.

Fixes processing#2402