Fix file path decoding
URLDecoder was being used for path decoding, even though it can't handle
RFC2396 encoding. This resulted in plus characters being removed and
possibly other weirdness.

See https://docs.oracle.com/javase/8/docs/api/java/net/URL.html
"The recommended way to manage the encoding and decoding of URLs is to
use URI, and to convert between these two classes using toURI() and
URI.toURL(). The URLEncoder and URLDecoder classes can also be used, but
only for HTML form encoding, which is not the same as the encoding
scheme defined in RFC2396."

Fixes #4417