Revert "Fix 3D on contemporary versions of Linux (fixes #5308)"
This reverts commit acee59918b4eceb749ef94d7a393403b223fcad3.

While this appears to have fixed GL for the AMD graphics users in #5308 it seems to have broke it for folks in #5476. It seems JOGL is abandoned these days, and I don't have the capacity to figure this out on my own. This patch reverts the "fix", so that it will continue to work on the systems that it traditionally worked on (and probably not on newer configuration not properly handled by JOGL).