Fix #444 by removing old regex parsing.
It looks like there's still some left over regex parsing to check for missing braces from before the new preproc. The new preproc will catch the error and testing makes it look like this old regex check is not triggering anyway outside of comments (which is an issue reported in #444).