java/preproc: Use class lookup to get the sketchname instead of hardcoding it
So far the sketchname was passed to the PApplet hardcoded as string in the generated
Java. While this approach is absolutely fine it is duplicate information.
For those who directly edit on the java lever (e.g. to get proper IDE support)
this means two points of edit when changing the sketch name.

Outside of the processing IDE there is no understanding of a sketch name only
a project name with a main class.

We therefore propose that the main class should be used as only source of truth
and the PApplet sketch name should be derived from it.
This also allows to extend the main class for multiple binaries from the same source.

Signed-off-by: Mimoja <git@mimoja.de>