Fix hash of keyEvent being added to pressedKeys.
When we press a letter, lets say `d` is key is char value `d` and hash calculated using keyCodea and key is added to pressed keys. But
when without release `d`, if we press `SHIFT` key, hash of shift key also being added to pressedKeys.
Now, when we release `d`, it is actually now `D` hence key value change and when we try to remove new hash, it doesnot exist in pressedKeys and the old hash of `d` still stays in pressedKeys which was the cause of keyPressed returning true.